added dinamic print values from registry vector and others
- refactor log output for detected sensors
- refactor cofirmation of each sensor detected
- added some second wire validations
- tested Adafruit bmp280 pull request

+++ src/Sensors.hpp
@@ -260,7 +260,7 @@ class Sensors {
 
     int getNextUnit();
 
-    uint32_t getUnitValue(UNIT unit);
+    float getUnitValue(UNIT unit);
 
    private:
     /// DHT library
